Move PrivateNoteElement to elements subfolder
