+++ app/packs/src/components/elements/PrivateNoteElement.js
